REFACTOR: Use globally-installed nw because installing the package for every project is wasteful (~100MB). FEATURE: Support bower packages in tags out-of-the-box.

+++ src/render.ls
@@ -8,11 +8,8 @@ _ = require('underscore')
 pkgconfig = require('./lib/pkg-config')
   
   
-macros = (PACKAGES, rootPath="") ->
+macros = (PACKAGES, rootPath="", more-tags={}) ->
   d = {}
-  mk-tag = (resource) ->
-    if resource is /\.js$/ then "<script src=\"#{resource}\"></script>"
-    else if resource is /\.css$/ then "<link rel=\"stylesheet\" text=\"text/css\" href=\"#{resource}\">"
   uri = (path, name) ->
     hack = if name is /^boiler/ then 1 else 0
     "#{rootPath}/#{path[hack]}/#name"
@@ -23,9 +20,29 @@ macros = (PACKAGES, rootPath="") ->
           ..[tag] = [mk-tag uri v.path, x for x in resources ? []].join "\n"
   d
 
+mk-tag = (resource) ->
+  if resource is /\.js$/ then """<script src="#{resource}"></script>"""
+  else if resource is /\.css$/ then "<link rel=\"stylesheet\" text=\"text/css\" href=\"#{resource}\">"
 
-projdir = -> window.projdir
-there = -> window.there
+
+bower-macros = (wd) -> {}
+  for dir in and-ancestors wd
+    bower = path.join dir, 'bower_components'
+    if fs.existsSync(bower) && fs.lstatSync(bower).isDirectory!
+      for module in fs.readdirSync(bower)
+        bower-json = path.join bower, module, 'bower.json'
+        if fs.existsSync(bower-json) && fs.lstatSync(bower-json).isFile!
+          try
+            main = JSON.parse fs.readFileSync bower-json, 'utf-8' .main
+            uri = (resource) -> path.relative(wd, path.join(bower, module, resource))
+            ..[module] = [mk-tag uri x for x in main].join "\n"
+          catch e
+            console.warn("Failed to read #module/bower.json (#bower-json)")
+
+and-ancestors = (dir) -> [dir]
+  while fs.realpathSync(dir = path.join(dir, '..')) != '/'
+    ..push dir
+
 
 Files =
   projdir: '.'
@@ -42,7 +59,7 @@ Files =
     
   rewriteFileSync: (filename, content) ->
     # This is done to refrain from tripping other fs watchers
-    if !fs.exists(filename) || fs.readFileSync(filename, 'utf-8') != content
+    if !(fs.exists(filename) && fs.readFileSync(filename, 'utf-8') == content)
       fs.writeFileSync(filename, content)
 
   Hash:
@@ -81,12 +98,13 @@ compile = (reload) ->
 
   in: ->
     settings = {interpolate: /<%=(.+?)[/%]>/g}
-    inputs = glob.sync("#{projdir}/**/*.in.html", {ignore: Files.ignore-patterns})
+    inputs = Files.find-all "*.in.html" .filter Files.Hash~is-dirty
+    #glob.sync("#{projdir}/**/*.in.html", {ignore: Files.ignore-patterns})
     inputs.map (input) ->
       output = input.replace(/\.in\.html$/, '.html')
       console.log "#{path.basename input} --> #{path.basename output}"
       rootDir = path.relative(path.dirname(output), incdir+'/public')
-      macros_ = macros(pkgconfig.PACKAGES, rootDir)
+      macros_ = macros(pkgconfig.PACKAGES, rootDir) <<< bower-macros(path.dirname(output))
       code = fs.readFileSync(input, 'utf-8')
       Files.rewriteFileSync(output, _.template(code, settings)(macros_))
